refactor functionality to usePagination hook
